+++ tests/test_parser/test_tf_to_spec.py
@@ -1,7 +1,7 @@
 import pytest
 import tensorflow as tf
 
-from vgslify.parsers.tf_parser import tf_to_spec
+from vgslify.parsers.tensorflow import tf_to_spec
 
 
 def test_input_layer():
