+++ vgslify/core/spec_parser.py
