+++ vgslify/model_parsers/__init__.py
@@ -1,0 +1,7 @@
+from .tensorflow import TensorFlowModelParser
+from .torch import TorchModelParser
+
+__all__ = [
+    "TensorFlowModelParser",
+    "TorchModelParser",
+]
