+++ vgslify/model_parsers/base.py
