--- vgslify/parsers/__init__.py
@@ -1,7 +1,0 @@
-from .tf_parser import TensorFlowModelParser
-from .torch_parser import TorchModelParser
-
-__all__ = [
-    "TensorFlowModelParser",
-    "TorchModelParser",
-]
